Added modern instructions for supporting iOS Silent Switch
Hey hey,

Currently doing an update to an old project and discovered that this library doesn't work for modern iOS. Took a while from there to figure out that AIR now has built in support for the Silent Switch. I've submitted this pull request to explain how to support the silent switch now that this library is deprecated!

Thanks!